Refactored namespace to be IntegrationTests since this better describes the testing methodology.
